little refactoring

+++ src/main/java/de/apnmt/organizationappointment/common/async/controller/AppointmentEventConsumer.java
@@ -2,10 +2,15 @@ package de.apnmt.organizationappointment.common.async.controller;
 
 import de.apnmt.common.controller.ApnmtEventConsumer;
 import de.apnmt.common.event.ApnmtEvent;
+import de.apnmt.common.event.ApnmtEventType;
 import de.apnmt.common.event.value.AppointmentEventDTO;
 import de.apnmt.organizationappointment.common.service.AppointmentService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class AppointmentEventConsumer implements ApnmtEventConsumer<AppointmentEventDTO> {
+
+    private final Logger log = LoggerFactory.getLogger(AppointmentEventConsumer.class);
 
     private final AppointmentService appointmentService;
 
@@ -15,13 +20,12 @@ public class AppointmentEventConsumer implements ApnmtEventConsumer<AppointmentE
 
     @Override
     public void receiveEvent(ApnmtEvent<AppointmentEventDTO> event) {
-        switch (event.getType()) {
-            case appointmentCreated:
-                this.appointmentService.save(event.getValue());
-                break;
-            case appointmentDeleted:
-                this.appointmentService.delete(event.getValue());
-                break;
+        if (event.getType() == ApnmtEventType.appointmentCreated) {
+            this.appointmentService.save(event.getValue());
+        } else if (event.getType() == ApnmtEventType.appointmentDeleted) {
+            this.appointmentService.delete(event.getValue());
+        } else {
+            log.error("Received wrong event type {}", event.getType());
         }
     }
 }
