+++ src/main/java/de/apnmt/organizationappointment/common/async/controller/OpeningHourEventConsumer.java
@@ -2,10 +2,15 @@ package de.apnmt.organizationappointment.common.async.controller;
 
 import de.apnmt.common.controller.ApnmtEventConsumer;
 import de.apnmt.common.event.ApnmtEvent;
+import de.apnmt.common.event.ApnmtEventType;
 import de.apnmt.common.event.value.OpeningHourEventDTO;
 import de.apnmt.organizationappointment.common.service.OpeningHourService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class OpeningHourEventConsumer implements ApnmtEventConsumer<OpeningHourEventDTO> {
+
+    private final Logger log = LoggerFactory.getLogger(OpeningHourEventConsumer.class);
 
     private final OpeningHourService openingHourService;
 
@@ -15,13 +20,12 @@ public class OpeningHourEventConsumer implements ApnmtEventConsumer<OpeningHourE
 
     @Override
     public void receiveEvent(ApnmtEvent<OpeningHourEventDTO> event) {
-        switch (event.getType()) {
-            case openingHourCreated:
-                this.openingHourService.save(event.getValue());
-                break;
-            case openingHourDeleted:
-                this.openingHourService.delete(event.getValue());
-                break;
+        if (event.getType() == ApnmtEventType.openingHourCreated) {
+            this.openingHourService.save(event.getValue());
+        } else if (event.getType() == ApnmtEventType.openingHourDeleted) {
+            this.openingHourService.delete(event.getValue());
+        } else {
+            log.error("Received wrong event type {}", event.getType());
         }
     }
 }
